Split and rename tests
